+++ Test/Cases/bug_include_directive.pas
@@ -1,0 +1,21 @@
+unit bug_include_directive;
+
+interface
+
+{$I jedi.inc}
+{$I extra.inc}
+
+uses
+  System.SysUtils;
+
+{$IFDEF SOMESYM}
+procedure P;
+{$ENDIF}
+
+implementation
+
+{$IFDEF SOMESYM}
+procedure P; begin end;
+{$ENDIF}
+
+end.